Generate sitemap at build time (and get rid of strange Next.js usage for XML generation)
