+++ app/blog/head.tsx
@@ -2,8 +2,6 @@ import { getImageUrl, getUrl } from "../../src/shared/urls";
 import { getTitle } from "../../src/shared/util";
 import preview from "../../src/shared/preview.png";
 
-export const homePageLastModified = "2022-04-30";
-
 const title = getTitle("Blog");
 const description =
   "Zurich-based software engineer \u00B7 Lead Architect at Zühlke \u00B7 I'm passionate about web development and UX. On this blog I explore working with technologies like React, Next.js, and TypeScript.";
