+++ app/blog/last-modified.ts
@@ -1,0 +1,3 @@
+const lastModified = "2022-04-30";
+
+export default lastModified;
