+++ package.json
@@ -3,7 +3,8 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "generate-sitemap": "ts-node -r dotenv/config ./scripts/generate-sitemap.ts",
+    "build": "next build && pnpm generate-sitemap",
     "analyze": "ANALYZE=true pnpm build",
     "start": "next start",
     "lint": "next lint",
@@ -46,6 +47,7 @@
     "@types/probe-image-size": "^7.0.1",
     "@types/react": "18.0.8",
     "autoprefixer": "^10.4.1",
+    "dotenv": "^16.0.3",
     "eslint": "8.6.0",
     "eslint-config-next": "13.0.4",
     "eslint-config-prettier": "^8.3.0",
@@ -54,6 +56,7 @@
     "prettier": "2.5.1",
     "prettier-plugin-tailwindcss": "^0.1.3",
     "tailwindcss": "^3.0.8",
+    "ts-node": "^10.9.1",
     "typescript": "4.5.4"
   },
   "version": "0.0.0"
